Use golangci-lint v2
Update to golangci-lint v2 and fix all of the lint issues.
To make it easier to debug the tests, log output has been surpressed by
default. This requires a user to provide a logger, otherwise logs will
now be discarded by default.

This also adds the testifylint to audit the way we're using testify.
As such, much of the incorrect usage of suite, assert, requires etc..
have also been fixed in this commit.

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>