fix: Integer overflow in allocation
CodeQL picked up an issue where we're allocating a slice based on the
size of 2 integers (size of db and client indicies). The addition
of these integers isn't bounds checked, so it's possible for this
to overflow or wraparound.

In practice this is likely not an issue since the integers we're
dealing with are small. Larger integers would cause a runtime
panic since the capacity would be out of range for a slice.

Either way, this adds a small function to take care of detecting
an overflow and propagates these errors back to the caller where
possible.

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>