test: Enable the integration tests to run on macOS
Uses OVS_VERSION as the single environment variable for specifying the OVS version.
This ensures the schemas and the version under test are consistent.

Replaces the hard-coded backup of 2.15.0 when OVS_VERSION is not set to use
"latest" instead. OVS_VERSION is set in the Makefile so this branch should not
be hit.

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>