ci: Test changes to images workflow
This adds a paths filter to the `push` event to only run the workflow
when files have changed. The same filter is used to run this on PR
events when the files have changed. On a PR run, we skip the
Docker login and image push since the credentials aren't available.
Worfklow_dispatch was added so the workflow can be manually run.

This should avoid needing to merge changes to main to test them
out.

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>