test: refactor common reconnect tests
The TestWithInactivity and TestWithReconnect logic is pretty much
identical. Move the common code into testReconnect. This also
removes the 2 ovsClients. Only a single client is used now.
A test may change the options at the beginning of the test.

Adds a TestWithFastInactivity variant with an aggressive timer
to try and reproduce #391.

This also fixes data races in the MonitorCondition tests.

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>