test: Reproduce deadlock in inactivity probes
This adds a test case that can reproduce the inactivity probe
deadlock that was reported in #391. It will deadlock reliably
therefore it's gated with a build tag. You must use:

```
go test -tags concurrency ./client -v -timeout 15s
```

This will run the test and dump a stacktrace when it times out

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>